refactor: rename caching package to cache
