+++ src/main/java/com/example/spinlog/statistics/scheduled/GenderStatisticsCacheRefreshScheduledService.java
@@ -2,7 +2,7 @@ package com.example.spinlog.statistics.scheduled;
 
 import com.example.spinlog.global.cache.CacheHashRepository;
 import com.example.spinlog.statistics.service.StatisticsPeriodManager;
-import com.example.spinlog.statistics.service.caching.GenderStatisticsCacheWriteService;
+import com.example.spinlog.statistics.service.cache.GenderStatisticsCacheWriteService;
 import com.example.spinlog.statistics.service.fetch.GenderStatisticsRepositoryFetchService;
 import com.example.spinlog.user.entity.Gender;
 import lombok.RequiredArgsConstructor;
