+++ src/main/java/com/example/spinlog/statistics/service/cache/GenderStatisticsCacheFallbackService.java
@@ -1,4 +1,4 @@
-package com.example.spinlog.statistics.service.caching;
+package com.example.spinlog.statistics.service.cache;
 
 import com.example.spinlog.article.entity.RegisterType;
 import com.example.spinlog.statistics.exception.InvalidCacheException;
