+++ src/main/java/com/example/spinlog/statistics/service/cache/StatisticsCacheSynchronizerAfterUserWrite.java
@@ -1,4 +1,4 @@
-package com.example.spinlog.statistics.service.caching;
+package com.example.spinlog.statistics.service.cache;
 
 import com.example.spinlog.statistics.service.StatisticsPeriodManager;
 import com.example.spinlog.statistics.service.fetch.GenderStatisticsRepositoryFetchService;
