+++ src/test/java/com/example/spinlog/integration/GenderStatisticsCacheUserWriteIntegrationTest.java
@@ -10,10 +10,7 @@ import com.example.spinlog.statistics.repository.dto.GenderDailyAmountSumDto;
 import com.example.spinlog.statistics.repository.dto.GenderEmotionAmountAverageDto;
 import com.example.spinlog.statistics.repository.dto.GenderSatisfactionAverageDto;
 import com.example.spinlog.integration.init.GenderStatisticsCacheSetupService;
-import com.example.spinlog.statistics.service.GenderStatisticsService;
-import com.example.spinlog.statistics.service.caching.GenderStatisticsCacheFallbackService;
-import com.example.spinlog.statistics.service.dto.GenderDailyAmountSumResponse;
-import com.example.spinlog.statistics.service.dto.GenderEmotionAmountAverageResponse;
+import com.example.spinlog.statistics.service.cache.GenderStatisticsCacheFallbackService;
 import com.example.spinlog.user.dto.request.UpdateUserRequestDto;
 import com.example.spinlog.user.entity.Gender;
 import com.example.spinlog.user.entity.User;
