+++ src/test/java/com/example/spinlog/statistics/scheduled/GenderStatisticsCacheRefreshScheduledServiceTest.java
@@ -4,7 +4,7 @@ import com.example.spinlog.article.entity.Emotion;
 import com.example.spinlog.statistics.repository.GenderStatisticsRepository;
 import com.example.spinlog.statistics.repository.dto.GenderEmotionAmountAverageDto;
 import com.example.spinlog.statistics.service.StatisticsPeriodManager;
-import com.example.spinlog.statistics.service.caching.GenderStatisticsCacheWriteService;
+import com.example.spinlog.statistics.service.cache.GenderStatisticsCacheWriteService;
 import com.example.spinlog.statistics.service.fetch.GenderStatisticsRepositoryFetchService;
 import com.example.spinlog.user.entity.Gender;
 import com.example.spinlog.util.CacheConfiguration;
