+++ src/test/java/com/example/spinlog/statistics/service/GenderStatisticsServiceTest.java
@@ -3,7 +3,7 @@ package com.example.spinlog.statistics.service;
 import com.example.spinlog.article.entity.Emotion;
 import com.example.spinlog.statistics.repository.GenderStatisticsRepository;
 import com.example.spinlog.statistics.repository.dto.*;
-import com.example.spinlog.statistics.service.caching.GenderStatisticsCacheFallbackService;
+import com.example.spinlog.statistics.service.cache.GenderStatisticsCacheFallbackService;
 import com.example.spinlog.statistics.service.dto.*;
 import com.example.spinlog.statistics.service.wordanalysis.WordExtractionService;
 import com.example.spinlog.user.entity.Gender;
