+++ src/test/java/com/example/spinlog/statistics/service/cache/GenderStatisticsCacheWriteServiceTest.java
@@ -1,4 +1,4 @@
-package com.example.spinlog.statistics.service.caching;
+package com.example.spinlog.statistics.service.cache;
 
 import com.example.spinlog.article.entity.RegisterType;
 import com.example.spinlog.global.cache.CacheHashRepository;
